+++ angular-16-client/src/app/components/tutorials-list/tutorials-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Tutorial } from 'src/app/models/tutorial.model';
 import { TutorialService } from 'src/app/services/tutorial.service';
+import { AuthService } from 'src/app/services/auth.service';
 
 @Component({
   selector: 'app-tutorials-list',
@@ -13,7 +14,7 @@ export class TutorialsListComponent {
   currentIndex = -1;
   title = '';
 
-  constructor(private tutorialService: TutorialService) {}
+  constructor(private tutorialService: TutorialService, public authService: AuthService) {}
 
   ngOnInit(): void {
     this.retrieveTutorials();
@@ -41,6 +42,10 @@ export class TutorialsListComponent {
   }
 
   removeAllTutorials(): void {
+    if (!this.authService.isAuthenticated()) {
+      return;
+    }
+
     this.tutorialService.deleteAll().subscribe({
       next: (res) => {
         console.log(res);
